Proxy tests running
