+++ integration/kube/fixtures.go
@@ -1,0 +1,120 @@
+package kube
+
+import (
+	"context"
+	"time"
+
+	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/integration/helpers"
+	"github.com/gravitational/teleport/lib/auth/native"
+	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/tlsca"
+	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/trace"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+)
+
+// For this test suite to work, the target Kubernetes cluster must have the
+// following RBAC objects configured:
+// https://github.com/gravitational/teleport/blob/master/fixtures/ci-teleport-rbac/ci-teleport.yaml
+const TestImpersonationGroup = "teleport-ci-test-group"
+
+type ProxyConfig struct {
+	T                   *helpers.TeleInstance
+	Username            string
+	KubeUsers           []string
+	KubeGroups          []string
+	Impersonation       *rest.ImpersonationConfig
+	RouteToCluster      string
+	CustomTLSServerName string
+	TargetAddress       utils.NetAddr
+}
+
+// ProxyClient returns kubernetes client using local teleport proxy
+func ProxyClient(cfg ProxyConfig) (*kubernetes.Clientset, *rest.Config, error) {
+	authServer := cfg.T.Process.GetAuthServer()
+	clusterName, err := authServer.GetClusterName()
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	// Fetch user info to get roles and max session TTL.
+	user, err := authServer.GetUser(cfg.Username, false)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	roles, err := services.FetchRoles(user.GetRoles(), authServer, user.GetTraits())
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	ttl := roles.AdjustSessionTTL(10 * time.Minute)
+
+	ca, err := authServer.GetCertAuthority(context.Background(), types.CertAuthID{
+		Type:       types.HostCA,
+		DomainName: clusterName.GetClusterName(),
+	}, true)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	caCert, signer, err := authServer.GetKeyStore().GetTLSCertAndSigner(ca)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	tlsCA, err := tlsca.FromCertAndSigner(caCert, signer)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	privPEM, _, err := native.GenerateKeyPair()
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	priv, err := tlsca.ParsePrivateKeyPEM(privPEM)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	id := tlsca.Identity{
+		Username:         cfg.Username,
+		Groups:           user.GetRoles(),
+		KubernetesUsers:  cfg.KubeUsers,
+		KubernetesGroups: cfg.KubeGroups,
+		RouteToCluster:   cfg.RouteToCluster,
+	}
+	subj, err := id.Subject()
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	cert, err := tlsCA.GenerateCertificate(tlsca.CertificateRequest{
+		Clock:     authServer.GetClock(),
+		PublicKey: priv.Public(),
+		Subject:   subj,
+		NotAfter:  authServer.GetClock().Now().Add(ttl),
+	})
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	tlsClientConfig := rest.TLSClientConfig{
+		CAData:     ca.GetActiveKeys().TLS[0].Cert,
+		CertData:   cert,
+		KeyData:    privPEM,
+		ServerName: cfg.CustomTLSServerName,
+	}
+	config := &rest.Config{
+		Host:            "https://" + cfg.T.Config.Proxy.Kube.ListenAddr.Addr,
+		TLSClientConfig: tlsClientConfig,
+	}
+	if !cfg.TargetAddress.IsEmpty() {
+		config.Host = "https://" + cfg.TargetAddress.Addr
+	}
+	if cfg.Impersonation != nil {
+		config.Impersonate = *cfg.Impersonation
+	}
+	client, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	return client, config, nil
+}
